requirements.txt
pymongo
beautifulsoup4
python-dateutil
pytest
